:test_tube: improving rules test e2e stability
