+++ packages/desktop-client/src/components/rules/ScheduleValue.tsx
@@ -5,8 +5,8 @@ import { describeSchedule } from 'loot-core/src/shared/schedules';
 import { type ScheduleEntity } from 'loot-core/src/types/models';
 
 import { usePayees } from '../../hooks/usePayees';
+import { useSchedules } from '../../hooks/useSchedules';
 
-import { SchedulesQuery } from './SchedulesQuery';
 import { Value } from './Value';
 
 type ScheduleValueProps = {
@@ -16,7 +16,7 @@ type ScheduleValueProps = {
 export function ScheduleValue({ value }: ScheduleValueProps) {
   const payees = usePayees();
   const byId = getPayeesById(payees);
-  const { data: schedules } = SchedulesQuery.useQuery();
+  const { data: schedules } = useSchedules();
 
   return (
     <Value
